linux-jitterentropy: trim the precompiled boundary
advice from atsec: jitterentropy-kcapi.c is not within the certified boundary.
don't precompile this, and undo the object shimming. it still needs to be patched
to use certified SHA3

drop function call shims from jitterentropy-testing.c. we've know for a while this didn't
need to be precompiled, just overlooked dropping the shims.