vms-test: gcp: retry instance launch on resource exhaustion
We have seen occasional CI failures where an instance launch for one variant
fails, but a launch of the same type in the same zone a little later succeeded
for another variant. To reduce spurious failures due to GCP capacity issues,
retry the launch for a total of 3 launch attempts.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>