fix(helm-operator-fips-1.33): finish withdrawal
This package was partially removed from our collection via the withdrawal in https://github.com/chainguard-dev/enterprise-packages/pull/31410, but the package build definition was left in place, so we've been repeatedly building the package and attempting to push it to our registry, which results in an error because the package had been explicitly deleted previously.

This PR finishes the removal by deleting the melange file.

Signed-off-by: Dan Luhring <dluhring@chainguard.dev>