vms-test: gcp: tweak launch retry mechanism (#629)
In last night's GCP CI run
(https://github.com/chainguard-dev/wolfi-vm/actions/runs/17253544217/job/48960811936),
the retry of instance launches helped a few times, but ultimately was
not sufficient. Tweak it a bit to be both more persistent and more
patient, before we may need to rework things a little to retry in a
different zone.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>