build-publish-aws: avoid misleading error output (#630)
Depending on timing, the catch-up mechanism for publishing to the AWS
Marketplace can cause harmless but misleading error output in the CI:

    /bin/sh: 3: echo: echo: I/O error
    /bin/sh: 3: echo: echo: I/O error
    /bin/sh: 3: echo: echo: I/O error
    /bin/sh: 3: echo: echo: I/O error
    /bin/sh: 3: echo: echo: I/O error
    make: *** [Makefile:79: list-aws-noneks] Error 1
Evaluating aws-agents (prod-tm2fmhgiqfrcs) for whether the group needs
to be published...

This is a broken pipe in disguise. The dash shell appears to report
EPIPE as an I/O error:

/bin/dash -c 'trap "" PIPE; for i in `seq 1 1000`; do echo $i; done |
head -1'

I suspect the GitHub Actions runner ignores SIGPIPE and make passes that
on. In any case, suppress the misleading error output by reading all
make output first. sort(1) does that job well.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>