vms-test: ssmsendcommand: wait for instance to ping SSM control plane
Amazon is not explicit in what constitutes a "ready" state for an
instance to receive commands from SSM/Systems Manager. The SSM Agent
test already waits for the instance to be registered with the control
plane, but that may not be sufficient as the CI still occasionally fails
this test. In an attempt to fix this, add the ping status as seen by
Systems Manager to the readiness checks.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>