vms-test: ssmsendcommand: wait for command to complete before exiting
The launch helper used by the SSM Agent test sends the command as soon
as possible and exits immediately after. The test harness then runs all
configured tests, and may actually run the verification part of the SSM
Agent test before SSM Agent on the instance had a chance to fetch and
execute the command. Wait for SSM to report the command has finished one
way or another before proceeding to run the tests.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>