Add ci command group
It felt like things were getting out of hand, so I've moved `stereo
lint` to `stereo ci lint` and added `unguarded` and `radius` commands as
well.

Currently, `lint` still just looks for duplicate package names. It's
fairly cheap to do that, and I expect we might want to add more cheap
things to it and/or split it out into a separate command that looks for
duplicate packages.

The `radius` command works against a local set of changes and will
attempt to solve every build environment for every package definition.
This takes a few seconds, but it will show you any builds that are
affected by any locally-built packages. We'll want to do something very
similar in CI but with the presubmit repos insetad of local stuff.

The `unguarded` command is similar to `radius`, but it shows you any
builds that, given your local source and the prod package repos,
currently pull in any packages we no longer build.

I've updated TRIAL.md to reflect this change.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>