build-publish-aws: enable compliance reporting for scheduled builds
The compliance reporting test originally was opt-in only, requiring a
manually dispatched workflow to run. This was done to not disturb the
nightly runs of the CI while shaking out bugs in the test. The test is
stable and works across a wide set of images now, so enable it in the
nightly runs now.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>