Use a shared cache for before/after in radius
We want to isolate the diff to just our local packages. This cache we
pass through to apko caches the results of fetching an APKINDEX, so we
don't observe any drift in the prod wolfi/extras/enterprise indexes when
we re-fetch them in the "after" locking.

I thought I had already wired this up, but obviously not!

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>