vms-test: make failure to install openscap package non-fatal
The compliance/reports test tries to install the openscap package if
apk-tools are present, but OpenSCAP is not. Up to now, it was reasonable
to assume failing to install the package means something is wrong with
the image or networking, and thus it was useful to fail the test. With
the addition of read-only images, the package may fail to install for
obvious reasons, hence just skip the test when openscap cannot be
installed.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>