passt - add package (#32964)
This package has some wierd tagging / version behavior.
1. the version is not valid apk version 2025_08_05
2. the tags include the commit hash 2025_08_05.309eefd

The packaging addresses version with just changing _ to '.'.
(2025_08_05 to 2025.08.05).  But the hash isn't easily
addressable

As a result, there isn't a sane way that we can make the updater
work.  As is, I expect the updater to make PRs that will require
a human to update the hash in the tag.

One thing to note is that git hash _is_ now supported in apk tools
via a '~'. See https://gitlab.alpinelinux.org/alpine/apk-tools/-/issues/10830
That is only in the 3.x branch now.  We could reasonably backport it
which would allow us to keep that information present in the version
(2025.08.05~309eefd).