withdrawn-packages: don't remove APK's from GCS
We have a number of APK's that are part of packages.cgr.dev/os but
are not present in apk.cgr.dev/chainguard-private; as such we can't use the
preferred restore process (which relies on the package having been
present in apk.cgr.dev).

The restore process for these older APK's will be to upload to
apk.cgr.dev using the APK from the GCS bucket.

Stop removing them as part of withdrawal to support this contingency.