builder - fix rootfs having contents of boot/efi and leave /dev dir. (#673)
The 'rm -rf "$root_dev/boot/efi/*"' was not expanding due to quotes.
This meant the rm was no doing anything and '-f' meant it would
fail silently.

The intention was fairly clearly to leave the directory but remove
the contents.

It is not possible to remove contents of the directory on the
filesystem without modifying the timestamp of the directory
and thus breaking reproducibility.

The change here is to extract the full tarball and then
move mount points out of it to filesystems/<base64 path>.
we take care to keep the directory's timestamp.

Then we populate filesystems from the "filesystems/<base64 path>"
directories.

One thing to note here is that this process is used to throw away
the contents of dev/ . That was being done before also.

The result is the following expected changes:
 * root filesystem should have dev/ directory entry
   (otherwise init cannot mount devfs on top of it)
 * boot/efi directory is present in the root filesystem but empty.