gc: withdraw APK's older than 12 months
As part of ongoing housekeeping for Wolfi, we will be removing APK's
older than 12 months. Older versions of APK's are assessed based on
usage within the archive, use in images and vms, use as build
dependencies and based on user telemetry reviews.

This is the September 10th Garbage Collection point.