build-publish-*: move scheduled workflows to 03:00 UTC
The scheduled build-publish-* workflows kicked off at 00:00 UTC, an
obvious, easy to remember, but also popular time of day. Move the
workflows a little to avoid the rush hour. This helps the stability of
the build workflows (e.g. failing due to octo-sts hitting GitHub rate
limits) of VMs and other projects.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>