Add bootc, bootupd and a linux package with an initramfs (#30878 )
This adds bootc and bootupd. 

bootupd is straight forward other than the fact that it was built from a branch.
bootupd updates bootloaders with new entries.  mainline has support for grub2,
but we need support for systemd-boot.

bootc is cut into two primary packages that have dependencies to listed to
make their use easy.  bootc-installer and bootc-system.  The runtime
system does not need to be able to install and thus we can do without
a lot of the runtime dependencies of `bootc install`.

Also adds a linux kernel subpackage that
1. contains an initramfs with composefs support 
2. contains kernel and initramfs in locations where bootc wants them to be (usr/lib/modules/<kver>)

This is _not_ the most ideal way to create an initramfs.  The bootc (and systemd)
that is present in the initramfs is the one that was current at time of the build.
One obvious problem with this approach is that a new version of bootc
then requires us to bump the kernel epoch so that we can get a new initramfs.
There are lots of other shortcomings on it.  We definitely need to revisit this.