linux - fix paths under /boot for boot-installed packages. (#33600)
* pipelines/kernel - fix paths under /boot for in boot-installed packages.

We were creating boot-installed packages that contained paths
like:

    boot/efi/chainguard/6.16.7-r1-qemu-generic/linux
    boot/efi/loader/entries/chainguard-6.16.7-r1-qemu-generic.conf

And as seen in the systemd pr listed below, we were placing
efi files in the filesystem under /boot/efi/EFI/bootx64.efi

systemd wants to mount the ESP partition directly to /boot (not
/boot/efi), and wants 'loader' to be a sibling of EFI.
In order for the apk database to be correct when the esp
partition is mounted at /boot, then we need to drop this
secondary 'efi'.

The fallout of this is that systemd would mount the ESP partition
to /efi rather than /boot.

With these changes in place, systemd will mount ESP partition to
/boot and apk database will be consistent with the filesystem view
at that point.

See also
  https://github.com/wolfi-dev/os/pull/66189

* linux-qemu-generic - adjust boot/efi path for bootc-boot-installed

* linux - bump to cause rebuild after pipeline/ change