amazon-ec2-net-utils: disable superfluous set-hostname-imds.service
The set-hostname-imds.service shipped in amazon-ec2-net-utils is a
recent feature addition that can set the hostname of a system based on
the value provided via the Instance Meta Data Service (IMDS). This is
unnecessary, as systemd-networkd already sets the hostname based on the
DHCP lease. Moreover, since querying the IMDS can fail while the
interface is being (re-)configured, having the service enabled can
actually be harmful and result in a degraded system state when it fails.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>