ghc-9.8: new package
This continues the Haskell bootstrap, this time building GHC from source
based on the stage0 package.  To do this in a reasonably reproducible
way, we need to pin dependencies, which Cabal doesn't make particularly
easy without a great deal of work to package many individual libraries.

The approach I've taken here is to construct a local mirror of sources
for the subset of upstream dependencies that we need, recording the
expected checksums for each of them.  Figuring out which dependencies
are needed is pretty hard work, and we certainly don't want to reinvent
Cabal, so I opted for a pipeline that can be used in a temporary local
build to record what Cabal does when given access to the internet and
dump the result into a JSON file (in a format similar to GHC's own
bootstrap plans), which another pipeline then uses to fetch dependencies
and then configure Cabal to use only what we've fetched.  These
pipelines can also be used for packages other than GHC.

My experience so far is that updating these dependencies is some work,
and it would certainly be nice to automate the process further; but as
long as the Haskell packages we're shipping are ones with a reasonably
low frequency of security updates, it should be sustainable.

Signed-off-by: Colin Watson <colin.watson@chainguard.dev>