enterprise: import a number of numpy-related packages (#34137)
Import a number of numpy-related packagers from wolfi os to enterprise.

This is part of a larger work to fix the dependency tree of modelmesh-runtime-adapter for the
kserve chainguard image.

We had a number of packages in the stack depending on different (and conflicting) versions
of numpy, mostly because misconfigure 'provides' directives, or in other cases, vanishing
virtual packages.

As a way to stabilize the dependency tree, we decided:
 * to depend on expliocit numpy version stream names
 * to target numpy 2.2 (enterprise), because 2.3 (os) loses py3.10 support.

Also, moving packages to enterprise and shrinking wolfi seems in line with other efforts
across chainguard.

Signed-off-by: Arturo Borrero Gonzalez <arturo.borrero@chainguard.dev>