airflow-2: unbreak python dependency hell
The pip install of the airflow wheel resulting in some weird dependency
resolution resulting in an ancient version of pyarrow being pulled into
the resolved set of modules; this version has no published wheels for
Python 3.12, so pip attempts to built them and fails as apache-arrow-dev
is not installed in the build environment.

The fix here is to avoid this dependency issue altogether; Apache
Airflow published constraint files that we can use for this purpose,
ensuring that the package is not impacted by some transient dependency
change on PyPi - this problem just gets more likely the older package
is.

This does pin all of the Python modules used but does result in a
reproducable build (so long as nothing actually gets removed from PyPi).