vms-test: restrict testing to regions to images in --region if set
Since a previous commit started creating AMIs in multiple regions, the
testing helpers more or less picked one at random. In practice, most
often the "wrong" region would win out and the testing harness failed to
find the infrastructure it expected.

For test-wolfi-vm-aws, restrict testing to the region specified via the
--region option (similar to --region for Azure or --zone for GCP) if
present. If not present, continue to test in all regions listed in the
awspub output.

For test-eks, ensure the image matching the AWS_REGION environment
variable is chosen. test-eks did not and continues to not support
testing in multiple regions at once.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>