vms-test: stop storing including redundant copies of the SCAP file
The SCAP security guide XML file used for compliance reporting is
already present on the host kicking off the test run, since it is there
that it was compiled into the test binary. Do not bother shipping
redundant copies of it back to the host, especially since the every
image will use the same SCAP security guide.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>