fix(tomcat-fips): Do not unset classpath (#34303)
When catalina.sh is invoked, the classpath is unset. This means that
BCFIPS is not available in the path at runtime

Unsetting the classpath doesn't really have any meaningful impact other
than possibly preventing users from specifying invalid paths. Other than
that, this is an anti-pattern

So patch catalina.sh to not unset it

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>