cabal-install-3.12: new package
This continues the Haskell bootstrap, building the `cabal` tool from
source based on its stage0 package.  This package is built using GHC 9.8
(since that's the newest stable version that cabal-install 3.10.3.0
supports) and can be used with GHC 9.10.

The bootstrap target is 3.16.0.0, but, as with GHC itself, we have to
work our way up to that one stable version series at a time.

Signed-off-by: Colin Watson <colin.watson@chainguard.dev>