fix(rancher-dashboard-2.12): Embed Harvester UI extension
This embeds the Harvester extension in the dashboard. The Harvester UI
extension is required by Rancher (as it provides the UI for Harvester in
Rancher). This addresses the UI failing to load

We use the version referenced in Rancher's CI directly to avoid an
incompatibility between Harvester's docker machine driver and the
interface

Also remove tests that aren't necessarily testing anything. Reasons:
- Serving the dashboard will fail if required assets are missing
- We know the package isn't empty as we explicity extract assets to the
  directory (and previously copied them)
- rancher-api-ui doesn't provide conflicting files as it doesn't write
  to the dashboard directory so installing the package during tests
  isn't necessary
- Validating softlinks is redundant; if creating the link fails, ln will
  fail verbosely

Serving the dashboard is sufficient

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>