configs: ecs: add cloud-ssh-keys-fetcher package back for now
Commit 6685182e ("ECS Image - Clean-up unnecessary packages (#745)")
removed a couple packages believed to not be necessary from the ECS
images, including cloud-ssh-keys-fetcher. The already present cloud-init
should be able to populate the authorized_keys file as needed. However,
as failing CI tests showed, it currently does so for the "chainguard"
user and not for "ec2-user".

Bring back cloud-ssh-keys-fetcher for now until we have changed our
cloud-init package to bring cloud-specific configuration for the default
user. [1] For fully immutable images using dm-verity, we are likely to
require cloud-ssh-keys-fetcher regardless, since cloud-init will not be
able to write the authorized_keys file.

[1]: https://github.com/chainguard-dev/wolfi-vm/issues/747

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>