vms-test: add test-ecs script for ECS cluster management
Add the new test-ecs script that manages ECS clusters for testing. Like
eksctl for test-eks, the weapon of choice for creating, updating,
deleting, and aggregating the status of a lot of AWS resources is
CloudFormation. Future work will see test-ecs extended to also manage
instances in clusters and eventually run tests on those.

Inspired by and based on work done by @smoser for test-eks.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>