Add --ignore file to unguarded command (#70)
Fixes #65

I'm also making lint and unguarded failures fatal so we can start
catching these sooner.

RJ took care of all the lint failures, and this adds a file for ignoring
known packages that we can no longer build but still depend on.

The unguarded command will also (non-fatally) warn if anything is on the
ignore list that doesn't need to be, so we can start to ratchet it
towards zero.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>