build-publish-aws: bump the lifetime of AWS creds used for testing
Bump the lifetime of the AWS credentials used for testing from the
default of 1 hour to 2 hours. The corresponding change in mono [1]
allows for credentials to live for up to 6 hours, but let's not fully
make use of that just now (as there's likely other efficiencies to be
had at that point, and limiting credential lifetime to what's necessary
is A Good Thing).

[1]: https://github.com/chainguard-dev/mono/pull/27547

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>