vms-test: test-ecs: add commands for instance management
Extend the test-ecs script to also manage instances for an ECS cluster,
specifically to launch instances and have them join the cluster. This
uses a cloud-init user data script to configure ECS Agent immediately
upon boot. Alternatively, instances could be configured to join the
cluster via SSM RunCommand, but this requires SSM Agent to be installed
and potentially lengthy interactions with the SSM service.

Provide a command to remove instances from a cluster and terminate them
for symmetry purposes. Since we don't want to reuse instances, there is
no need to explicitly deregister them from the cluster. The ECS control
plane is notified of the instance termination and cleans up any tasks
assigned to the instance.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>