Makefile: rebuild builder initrd if selected builder changes
So far, an explicit "make builder" is required to rebuild the initrd
containing the selected builder from iac/. For example, in

  1. make disk-aws-base
  2. make disk-aws-base BUILDER=builder-immutable

the invocation in 2. will have make find there is nothing to be done.
Start tracking the chosen builder (and not just any builder) initrd as a
dependency of an image build.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>