Makefile: automatically select a builder for image builds
We have multiple builders defined in iac/ for differing needs. The
situation will hopefully change soon to unify them, but for the time
being it will be helpful to automatically select an appropriate builder
for image builds when none is explicitly provided via the BUILDER
variable--otherwise, CI workflows will have to take care to rebuild
builders and explicitly list the images to be built with each of them
(there is an example of that in the Raspberry Pi workflow already).

An explicitly requested builder--via the BUILDER= command line argument
or a BUILDER environment variable--will take precedence over an
automatically selected one. If no rule for auto-selection fits, the
default builder "builder" will be used as before.

Behavior should not change for now: There are no image build definitions
that fit the "hardened" rule yet, and the "rpi" build definitions are
used with an explicitly defined BUILDER in the CI workflow. Eventually,
both "immutable" images and "debug" images should be able to become
additional rules in "select_builder".

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>