cli: outdated and check
This PR is a little too large but I wanted to evolve these commands
together.

We've talked about `stereo outdated` a bit before, which helps surface
any packages that are accidentally pinned to old versions. This comes
with its owns exceptions list, "outdated.txt".

For performance reasons, I've also lumped three of these into a new
`stereo check` command, which does some expensive stuff up front and
uses that to run `impact`, `unguarded`, and `outdated` against a set of
new packages.

This command will take a `--presubmit` flag and reconstruct the
necessary extra_repos to evaluate everything we want to know about a PR.

In mono, I'm going to set up something that just runs `stereo check`
with the right parameters. We'll pipe the output directly into a
markdown file so we don't have to update mono in order to update checks.

For staging, we're going to do some hack things. I'll plumb some
config-driven stuff into this later so we don't have to hardcode it
(which we will need to do anyway for eco-2-28).

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>