unguarded: Print any packages we fail to lock (#74)
We were just ignoring these before, which confused me when I saw
chromium suddenly appear as a failure. This is because yesterday we were
ignoring chromium because we couldn't lock it at all. Today we can lock
it, but it's pulling in unguarded deps.

I'm torn on making these failures fatal with an allowlist. I think it
will be too noisy right now, but we could make it fatal to help ratchet
in the future.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>