vms-test: runner: keep test binaries in /home instead of /tmp
On hardened images, /tmp is mounted noexec. The test runners
consequently fail to execute any tests there:

    2025/10/02 19:16:20 Connecting to VM at 44.197.231.1
    2025/10/02 19:16:21 failed to run systemd on remote host: command error: Process exited with status 1
    Output: sudo-rs: cannot execute '/tmp/systemd': Permission denied (os error 13)

Store the tests in $HOME instead, since /home is not mounted noexec on
hardened images. As of systemd-ephemeral-0.0.3, /home is an ephemeral
overlayfs mount on immutable images, so the location should not pose any
problems there either.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>