vms-test: test-ecs: cache CloudFormation DescribeStacks responses
The cluster_resources function retrieves relevant outputs of a
CloudFormation stack, but performs one API call per function call. Make
cluster_resources cache the response of that API call since the state of
the CloudFormation stack is not going to change during execution of the
test. This allows cheap use of the cluster_resources function without
having to plumb through its results to all functions that may care about
them.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>