Rename 'test' rule to 'test-gotest', use make to invoke go test in c-i (#786)
Rename the 'test' target to 'test-gotest' to be more indicative of what
it does.

In c-i, run the go tests with 'make' rather than a copy of the 'go test'
command line. That does 2 things:
1. less repeating ourselves
2. allow a PR to change what happens in 'test-gotest' or what it depends
on.

Without this, c-i runs main's copy of .github/workflows/go-auth-test, so
you can't test that the changes worked in presubmit.