add setup-melange to presubmit-reproducible (#790)
Similar to #788, The pr in #749 needs melange.
In order for its presubmit tests to be meaningful, we need to add
melange on mainline.

We only needed this change in presubmit-reproducible and
presubmit-go-auth-test because the other disk builder pre-submit jobs
use the `build-disks` action here, which is read from the submitted
PR and not from mainline.