Use a single builder, move installer to cgr-install (#749)
The immediate goal here is to move the installer code out of yaml-bash
inside of iac/build-*.yaml and into a repo of its own where we can
remove code duplication and ultimately get out of bash entirely.

apkoaas is extended with a '--kernel-cmdline-append' flag which adds the
argument to the kernel command line. This allows us to pass dynamic
build-time arguments through the iac/builder code.

tools/get-install-opts is invoked via Makefile during the disk target.
It is responsible for deciding which options are passed to
'--kernel-cmdline-append'. It is crappy shell script right now.

I personally think the way forward is to have information like 'rpi' or
'immutable' and other install information in the image build config, or
at very least a declarative config file. For now, we heuristically
decide a build is 'rpi' or 'immutable' based on the config name.

Note: the immutable, rpi and cis builders are still being used via the
mechanism that Markus added in #769.