ct-manylinux-2.28: Stop statically linking all of libstdc++
Statically linking libstdc++ can cause scenarios where multiple
libstdc++ runtimes are loaded in a single application, which
can cause crashes and other undefined behavior.

Apply the patches from AlmaLinux's gcc-toolset packages that generate static
"libstdc++_nonshared" libraries containing the symbol delta between GCC 8's
libstdc++.so and the version of GCC we're building. Binaries will now
dynamically link against the system's libstdc++.so.6, and we use a linker
script to statically link in any of the libstdc++_nonshared symbols they
need.

Build an oldlibstdcxx-2.28 package that provides a library to link against at
build time, and to serve as a reference for ABI testing.

We add -D_GLIBCXX_ASSERTIONS to our libstdc++ build flags because that impacts
the internal libstdc++ ABI and is what the compat layer expects. We may want
to consider this for our wolfi compilers as well.

A handful of symbols remain that don't match AlmaLinux's compat layer. I've
patched out a couple that are in AlmaLinux's compat layer but not present
in our build (otherwise the test fails to build the nonshared .so). The
test/oldlibstdc++ pipeline reports a few more that we don't have compat
symbols for, but that test is currently non-fatal. I'll continue to narrow
those down, and update the test to make it fail if any new ones appear. But
this is enough to build PyTorch, which we urgently need to ship.

Ideally the gcc-toolset-* patches should live in their own packages so that
our update bot can notice new tags. I tried to do that, but the versioning
is not APK compatible and will need significant munging, so I've held off on
that for now.

Fixes: https://github.com/chainguard-dev/internal-dev/issues/20536
Fixes: https://github.com/chainguard-dev/internal-dev/issues/20924

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>