build-publish-aws: integrate test-ecs into the CI workflow
Add a new test-ecs-container-instance GitHub action similar to (and
originally copied from) the existing test-eks-node action. Fun fact:
It's a Kubernetes worker node, but an ECS Container Instance.

This moves testing of ECS images from the chainguard-vms AWS account to
the chainguard-test-vms account. The relevant ECS role policy will be
attached to the post-submit role of this account (PR is waiting to be
merged).

The ECS testing steps are supposed to not break the workflow in case
testing fails. This both to test the waters and to allow merging
independent of the PR adding the required permissions to the AWS role.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>