linux: fix a memory leak in the jitterentropy kernel
The jent_hash_time function of the jitterentropy kernel leaks the desc
sha3_state object it is allocating on the heap. This memory is
unreclaimable, causing memory pressure, soft lock-ups, and eventually an
unresponsive host.

Move the sha3_state back onto the stack. That avoids the overhead of the
slab allocator and sidesteps potential memory leaks. The struct fits on
the stack as it is 356 bytes in size; less than the 360+ bytes taken up
by the expanded SHASH_DESC_ON_STACK removed with the reproducible
jitter-entropy patch.

See: https://github.com/chainguard-dev/wolfi-vm/issues/838

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>