linux: fix mangled-package-version variable transform
A recent Melange change [1] requires that all variable transformation
regular expressions match. The Linux configurations so far relied on the
regular expression not having to match to deal with a peculiarity of new
mainline kernel releases: The kernel version will be 6.12.0, but the Git
tag will be v6.12 (without the patch level).

Modify the variable transform regular expression and substitution
pattern to always match, and yet produce the desired outcome.

Intentionally do not bump the epoch. There is no need to rebuild the
packages. This just enables future rebuilds.

[1]: https://github.com/chainguard-dev/melange/pull/2114

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>