generic-image-upload: shellcheck fixes (#857)
Shellcheck has two valid findings

```
In generic-image-upload line 102:
        --tier "$tier" \
        ^----^ SC2215 (warning): This flag is used as a command name. Bad line break or missing [ .. ]?

In generic-image-upload line 354:
            vmkd_flat_gcs_path=$(upload_gcs_file "$vmdk_flat_path" "$vmdk_flat_gcs_path" "$vmdk_flat_filename") || return 1
            ^----------------^ SC2034 (warning): vmkd_flat_gcs_path appears unused. Verify use (or export if used externally).

For more information:
  https://www.shellcheck.net/wiki/SC2034 -- vmkd_flat_gcs_path appears unused...
  https://www.shellcheck.net/wiki/SC2215 -- This flag is used as a command na...
```

Re-errage --tier such that it is passed to the command line. Also
rearrage redirect to a new line, such that it is easy to add any
additional flags in the future without errors.

Fix typo in variable name for the flat vmdk, as it is spelled
correctly on the next line.