feat(python): Add pipelines for consuming Chainguard Libraries (#36174)
* feat(python): Add pipelines for consuming Chainguard Libraries

This adds a few pipelines for both authentication with our indexes, and for building packages that consume Chainguard Libraries:

- py/install: Leverages uv to install individual Python packages from
  Chainguard Libraries. Used exclusively for individual Python packages
- auth/uv: Configures uv to pull Chainguard Libraries

An auth pipeline has not been added for pip as we plan to start by
dogfooding individual Python packages gradually, and work from there.

This change also migrates our py3-fastapi packages to Chainguard
Libraries. fastapi is only used across a couple images and updated
frequently so it makes sense to start here. It's also Python-only

Todo in followup PRs:
- Support falling back to source builds without switching pipelines
- In those instances, also support bootstraps

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>

* chore(py3-fastapi): Use Chainguard Libraries

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>

* chore(auth/uv): no-index-fallback -> allow-unsafe-indexes

Don't use a double negative when naming toggles

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>

* chore(py/install): Enforce use of Chainguard Libraries

We'd expect every individual library we package to use our libraries and
no other index so we don't need to accomodate fallback

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>

* chore: Rename py/install to py/chainguard-lib-install

Reinforces that this pipeline is for installing individual libraries
that come from us

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>

* chore(py3-fastapi): Remove comments

Let git history be the source of truth

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>

* chore(auth/uv): Mark uv as needed

There is no scenario ever wherein you would use this pipeline and not uv

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>

* chore(auth/uv): Clarify cg libraries are used with allow-unsafe-indexes

allow-unsafe-indexes provides fallback to other Python indexes in the
absence of a library in cg libraries

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>

---------

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>