Fix pre-commit unittests andgo-auth-test, Enable kvm/vhost (#899)
`make test-gotest` was broken in GHA, and `make -C vms-test unittest` was broken everywhere.

unittest is fixed is updated to account for vhost changes and not use random CID in test.
the test-gotest is fixed in GHA by giving it access to vhost.

Also here, we actually give x86_64 GHA runners access to kvm, so they now use kvm.
You can see that in logs where we're using `-machine q35,accel=kvm`, so there might
be an improvement in performance there.