Add support for calling qemu with qcow snapshots and use in run- and test. (#873)
qemu `-snapshot` is documented to not work with arguments used for
`-blockdev`.
Since we are using -blockdev, provide the Snapshotify function that just
goes
through and makes qcow2 snapshots the way that qemu would if it worked.

Subsequent test runs would then be stateful from previous runs. This is
not what we would want. Each test should start with a fresh disk.raw and
firwmare
files.