github: set aws max-concurrency to 2
It is not clear if 50 simultanious ImportSnapshot jobs is a temporary
limit due to recent us-east-1 outage, or if it is a regular limit.

If all AWS jobs start in parallel we exceed 50 jobs. For now we have
less that 25 images in every arch-category of aws images (main, ecs,
eks), thus set max-concurrency to 2 such that no two aws matrix jobs
reach the 50 concurrent ebs snapshot import limits.