Remove duplicate qemu code from vms-test, use pkg/utils (#933)
The changes yesterday to put this all under the same go module allow us
to reduce the copy-pasta that was here.

generateArmCommand and generateAmd64Command are largely the same in
vms-test as they are in pkg/, but not exactly. So I've left them for
now. Also left untouched is 'GetCommand' in vms-test that is basically
utils.GenerateQEMUCommand.