linux: update configs for 6.17
Rerun oldconfigs for x86_64 and aarch64, as our current configs are
based of 6.16 kernel.

Most of the churn is just upstream refactoring renaming things.

Most new niche devices support are optional, and not enabled.

DEVTMPFS_MOUNT was not enabled aarch64, not sure how or why, as it was
on on x86_64.

These cleanups ahead of making updates to other configs.