var/run - handle whatever you find (#957)
The previous solution here would create /var/run/run if the /var/run was
a directory.

The over-engineered solution here should report reasonably well what it
did and make sure that /var/run is a symlink to ../run.

One exception to that is that I'm choosing to leave an existing absolute
symlink - although I don't know why that would be there.